More work on codeGen, functions next.
* also added unique name generator

+++ source/EwcTypes.h
@@ -165,11 +165,11 @@ static_assert(sizeof(s64) == 8, "wha");
 static_assert(sizeof(s32) == 4, "wha");
 
 inline s32 S32Coerce(s64 n)		{ s32 nRet = (s32)n;	EWC_ASSERT((s64)nRet == n, "S32Coerce failure"); return nRet; }
-inline s16 S16Coerce(s32 n)		{ s16 nRet = (s16)n;	EWC_ASSERT((s32)nRet == n, "S16Coerce failure"); return nRet; }
-inline s8 S8Coerce(s32 n)		{ s8 nRet = (s8)n;		EWC_ASSERT((s32)nRet == n, "S8Coerce failure");  return nRet; }
+inline s16 S16Coerce(s64 n)		{ s16 nRet = (s16)n;	EWC_ASSERT((s64)nRet == n, "S16Coerce failure"); return nRet; }
+inline s8 S8Coerce(s64 n)		{ s8 nRet = (s8)n;		EWC_ASSERT((s64)nRet == n, "S8Coerce failure");  return nRet; }
 inline u32 U32Coerce(u64 n)		{ u32 nRet = (u32)n;	EWC_ASSERT((u64)nRet == n, "u32Coerce failure"); return nRet; }
-inline u16 U16Coerce(u32 n)		{ u16 nRet = (u16)n;	EWC_ASSERT((u32)nRet == n, "u16Coerce failure"); return nRet; }
-inline u8 U8Coerce(u32 n)		{ u8 nRet = (u8)n;		EWC_ASSERT((u32)nRet == n, "u8Coerce failure");  return nRet; }
+inline u16 U16Coerce(u64 n)		{ u16 nRet = (u16)n;	EWC_ASSERT((u64)nRet == n, "u16Coerce failure"); return nRet; }
+inline u8 U8Coerce(u64 n)		{ u8 nRet = (u8)n;		EWC_ASSERT((u64)nRet == n, "u8Coerce failure");  return nRet; }
  
 // type traits
 template <typename T> struct SStripConst				{ typedef T Type;	enum { F_STRIPPED= false }; };
